% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjustedHammingDist}
\alias{adjustedHammingDist}
\title{Auxiliary function used by MLeNN. Computes the Hamming Distance between two instances}
\usage{
adjustedHammingDist(x, y, D)
}
\arguments{
\item{x}{Index of sample 1}

\item{y}{Index of sample 2}

\item{D}{mld \code{mldr} object in which the instances are located}
}
\value{
The Hamming Distance between the instances
}
\description{
Auxiliary function used by MLeNN. Computes the Hamming Distance between two instances
}
