% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSMOTE.R
\name{calculateDistances}
\alias{calculateDistances}
\title{Auxiliary function used by MLSMOTE. Calculates the distances between an instance and the ones with the same label}
\usage{
calculateDistances(sample, minBag, label)
}
\arguments{
\item{sample}{Sample whose distances to other samples with the same label we want to know}

\item{minBag}{Instances with a determined label}

\item{label}{Minoritary label we want to balance}
}
\value{
A list with the distances to the rest of samples with the same label
}
\description{
Auxiliary function used by MLSMOTE. Calculates the distances between an instance and the ones with the same label
}
\examples{
calculateDistances(25,c(75,65,89,23),300)
}
