% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateTableVDM}
\alias{calculateTableVDM}
\title{Auxiliary function used to calculate an auxiliary table to make VDM calculation faster}
\usage{
calculateTableVDM(D)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}
}
\value{
A dataframe with tables, useful for VDM calculation
}
\description{
Auxiliary function used to calculate an auxiliary table to make VDM calculation faster
}
