% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateInstanceMLSOL}
\alias{generateInstanceMLSOL}
\title{Auxiliary function used by MLSOL. Creates a synthetic sample based on two other samples, taking into account their types}
\usage{
generateInstanceMLSOL(seedInstance, refNeigh, t, D)
}
\arguments{
\item{seedInstance}{Index of the sample we are using as "template"}

\item{refNeigh}{Index of the reference neighbor}

\item{t}{types of the instances}

\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}
}
\value{
A synthetic sample derived from the one passed as a parameter and its neighbors
}
\description{
Auxiliary function used by MLSOL. Creates a synthetic sample based on two other samples, taking into account their types
}
