% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllNeighbors2}
\alias{getAllNeighbors2}
\title{Auxiliary function used by MLeNN and MLTL. Gets the kNN of every instance in a dataset, when compared to some of the rest}
\usage{
getAllNeighbors2(neighbors, d, k)
}
\arguments{
\item{neighbors}{Structure with all the neighbors in the dataset, regardless of which ones to be compared}

\item{d}{Vector with the instances of the dataset which are going to be compared}

\item{k}{Number of neighbors to be retrieved}
}
\value{
A list of vectors with the indexes of the neighbors for each instance
}
\description{
Auxiliary function used by MLeNN and MLTL. Gets the kNN of every instance in a dataset, when compared to some of the rest
}
