% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllReverseNeighbors}
\alias{getAllReverseNeighbors}
\title{Auxiliary function used by MLUL. For each instance in the dataset, given the neighbors structure, we compute its reverse nearest neighbors}
\usage{
getAllReverseNeighbors(d, neighbors, k)
}
\arguments{
\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}

\item{neighbors}{Structure with the neighbors of every instance in the dataset}

\item{k}{Number of neighbors to be considered}
}
\value{
A list of vectors with the indexes of the reverse nearest neighbors of every instance in the dataset
}
\description{
Auxiliary function used by MLUL. For each instance in the dataset, given the neighbors structure, we compute its reverse nearest neighbors
}
