% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getC}
\alias{getC}
\title{Auxiliary function used by MLSOL and MLUL. For each instance in the dataset, we compute, for each label, the proportion of neighbors having an opposite class with respect to the proper instance}
\usage{
getC(D, d, neighbors, k)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}

\item{neighbors}{Structure with the neighbors of every instance in the dataset}

\item{k}{Number of neighbors taken into account for each instance}
}
\value{
A structure with the proportion of neighbors having an opposite class with respect to an instance and label
}
\description{
Auxiliary function used by MLSOL and MLUL. For each instance in the dataset, we compute, for each label, the proportion of neighbors having an opposite class with respect to the proper instance
}
