% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNN}
\alias{getNN}
\title{Auxiliary function used to compute the neighbors of an instance}
\usage{
getNN(sample, rest, label, D, tableVDM = NULL)
}
\arguments{
\item{sample}{Index of the sample whose neighbors we want to know}

\item{rest}{Indexes of the samples among which we will search}

\item{label}{Label that must be active, in order to calculate the distances}

\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{tableVDM}{Dataframe object containing previous calculations for faster processing. If it is empty, the algorithm will be slower}
}
\value{
A vector with the indexes inside rest of the neighbors
}
\description{
Auxiliary function used to compute the neighbors of an instance
}
