% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNumCores}
\alias{getNumCores}
\title{Get the number of cores available for parallel computing}
\usage{
getNumCores()
}
\value{
The number of cores available for parallel computing
}
\description{
Get the number of cores available for parallel computing
}
\examples{
getNumCores()

}
