% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getS}
\alias{getS}
\title{Auxiliary function used by MLSOL and MLUL. For non outlier instances, it aggregates the values of C, taking into account the global class imbalance}
\usage{
getS(D, d, C, minoritary)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}

\item{C}{Structure with the proportion of neighbors having an opposite class with respect to an instance and label}

\item{minoritary}{Vector with the minoritary class of each label (normally, 1)}
}
\value{
A structure with the proportion of neighbors having an opposite class with respect to an instance and label, normalized by the global class imbalance
}
\description{
Auxiliary function used by MLSOL and MLUL. For non outlier instances, it aggregates the values of C, taking into account the global class imbalance
}
