% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getU}
\alias{getU}
\title{Auxiliary function used by MLUL. It computes the influence of each instance with respect to its reverse neighbors}
\usage{
getU(D, d, rNeighbors, S)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}

\item{rNeighbors}{Structure with the reverse nearest neighbors of each instance of the dataset}

\item{S}{Structure with the proportion of neighbors having an opposite class with respect to an instance and label, normalized by the global class imbalance}
}
\value{
A list of values of influence for each instance with respect to its reverse neighbors
}
\description{
Auxiliary function used by MLUL. It computes the influence of each instance with respect to its reverse neighbors
}
