% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getV}
\alias{getV}
\title{Auxiliary function used by MLUL. It calculates, for each instance, how important it is in the dataset}
\usage{
getV(d, w, u)
}
\arguments{
\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}

\item{w}{List of weights for each instance}

\item{u}{List of influences in reverse neighbors for each instance}
}
\value{
A list with the values of importance of each instance in the dataset
}
\description{
Auxiliary function used by MLUL. It calculates, for each instance, how important it is in the dataset
}
\examples{
\dontrun{
library(mldr)
getV(rownames(bibtex), w, u)
}
}
