% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_ROS.R
\name{LP_ROS}
\alias{LP_ROS}
\title{Randomly clones instances with minoritary labelsets}
\source{
Francisco Charte, Antonio J. Rivera, María J. del Jesus, and Francisco Herrera. Addressing imbalance in multilabel classification: Measures and random resampling algorithms. Neurocomputing, 163:3–16, 2015. ISSN 0925-2312. doi:https://doi.org/10.1016/j.neucom.2014.08.091
}
\usage{
LP_ROS(D, P)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{P}{Percentage in which the original dataset is increased}
}
\value{
An mldr object containing the preprocessed multilabel dataset
}
\description{
This function implements the LP-ROS algorithm. It is a preprocessing algorithm for imbalanced multilabel datasets,
whose aim is to identify instances with minoritary labels, and randomly clone them.
}
\examples{
library(mldr)
LP_ROS(bibtex, 25)
}
