% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_RUS.R
\name{LP_RUS}
\alias{LP_RUS}
\title{Title}
\usage{
LP_RUS(D, P)
}
\arguments{
\item{D}{Original multilabel dataset}

\item{P}{Percentage in which the original dataset is reduced}
}
\value{
Processed multilabel dataset, more balanced
}
\description{
Title
}
\examples{
LP_RUS(bibtex, 25)
}
