% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_RkNNOS.R
\name{ML_RkNNOS}
\alias{ML_RkNNOS}
\title{Reverse-nearest neighborhood based oversampling for imbalanced, multi-label datasets}
\source{
Sadhukhan, P., & Palit, S. (2019). Reverse-nearest neighborhood based oversampling for imbalanced, multi-label datasets. Pattern Recognition Letters, 125, 813-820
}
\usage{
ML_RkNNOS(D, k)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{k}{Number of neighbors to be considered when creating a synthetic instance}
}
\value{
A mld object containing the preprocessed multilabel dataset
}
\description{
This function implements an algorithm that uses the concept of reverse nearest neighbors, in order to create new instances for each label. Then, several radial SVMs, one for each label, are trained in order to predict each label of the synthetic instances.
}
\examples{
\dontrun{
library(mldr)
ML_RkNNOS(bibtex, 3)
}
}
