% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_ROS.R
\name{ML_ROS}
\alias{ML_ROS}
\title{Title}
\usage{
ML_ROS(D, P)
}
\arguments{
\item{D}{Original multilabel dataset}

\item{P}{Percentage in which the original dataset is increased}
}
\value{
Processed multilabel dataset, more balanced
}
\description{
Title
}
\examples{
ML_ROS(bibtex, 25)
}
