% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_RUS.R
\name{ML_RUS}
\alias{ML_RUS}
\title{Title}
\usage{
ML_RUS(D, P)
}
\arguments{
\item{D}{Original multilabel dataset}

\item{P}{Percentage in which the original dataset is reduced}
}
\value{
Processed multilabel dataset, more balanced
}
\description{
Title
}
\examples{
ML_RUS(bibtex, 25)
}
