% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLRkNNOS.R
\name{MLRkNNOS}
\alias{MLRkNNOS}
\title{Reverse-nearest neighborhood based oversampling for imbalanced, multi-label datasets}
\source{
Sadhukhan, P., & Palit, S. (2019). Reverse-nearest neighborhood based oversampling for imbalanced, multi-label datasets. Pattern Recognition Letters, 125, 813-820
}
\usage{
MLRkNNOS(D, k, tableVDM = NULL)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{k}{Number of neighbors to be considered when creating a synthetic instance}

\item{tableVDM}{Dataframe object containing previous calculations for faster processing. If it is empty, the algorithm will be slower}
}
\value{
A mld object containing the preprocessed multilabel dataset
}
\description{
This function implements an algorithm that uses the concept of reverse nearest neighbors, in order to create new instances for each label. Then, several radial SVMs, one for each label, are trained in order to predict each label of the synthetic instances.
}
