% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSMOTE.R
\name{MLSMOTE}
\alias{MLSMOTE}
\title{Synthetic oversampling of multilabel instances (MLSMOTE)}
\source{
Francisco Charte, Antonio J. Rivera, María J. del Jesus, and Francisco Herrera. MLSMOTE: Approaching imbalanced multilabel learning through synthetic instance generation. Knowledge-Based Systems, 39:385–397, 2015. ISSN 0950-7051. doi:https://doi.org/10.1016/j.knosys.2015.07.019
}
\usage{
MLSMOTE(D, k)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{k}{Number of neighbors to be considered when creating a synthetic instance}
}
\value{
A mld object containing the preprocessed multilabel dataset
}
\description{
This function implements the MLSMOTE algorithm. It is a preprocessing algorithm for imbalanced multilabel datasets,
whose aim is to identify instances with minoritary labels, and generate synthetic instances based on their neighbor instances.
}
\examples{
\dontrun{
library(mldr)
MLSMOTE(bibtex, 3)
}
}
