% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSOL.R
\name{MLSOL}
\alias{MLSOL}
\title{Multi-label oversampling based on local label imbalance (MLSOL)}
\source{
Liu, B., Blekas, K., & Tsoumakas, G. (2022). Multi-label sampling based on local label imbalance. Pattern Recognition, 122, 108294.
}
\usage{
MLSOL(D, P, k)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{P}{Percentage in which the original dataset is increased}

\item{k}{Number of neighbors to be considered when computing the neighbors of an instance}
}
\value{
A mld object containing the preprocessed multilabel dataset
}
\description{
This function implements the MLSOL algorithm. It is a preprocessing algorithm for imbalanced multilabel datasets,
whose aim is to identify instances with minoritary labels, and generate synthetic instances based on their neighbor instances.
}
\examples{
\dontrun{
library(mldr)
MLSOL(bibtex, 3)
}
}
