% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLTL.R
\name{MLTL}
\alias{MLTL}
\title{Multilabel approach for the Tomek Link undersampling algorithm (MLTL)}
\source{
Rodolfo M. Pereira, Yandre M.G. Costa, and Carlos N. Silla Jr. MLTL: A multi-label approach for the Tomek Link undersampling algorithm. Neurocomputing, 383:95–105, 2020. ISSN 0925-2312. doi:https://doi.org/10.1016/j.neucom.2019.11.076
}
\usage{
MLTL(D, TH, neighbors = NULL)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{TH}{threshold for the Hamming Distance in order to consider an instance different to another one.}

\item{neighbors}{Structure with instances and neighbors. If it is empty, it will be calculated by the function}
}
\value{
An mldr object containing the preprocessed multilabel dataset
}
\description{
This function implements the MLTL algorithm. It is a preprocessing algorithm for imbalanced multilabel datasets, whose aim is to identify
tomek links (majoritary instances with a very different neighbor), and remove them. It's like MLeNN, with the number of neighbors being 1.
}
\examples{
\dontrun{
library(mldr)
MLTL(bibtex, 0.5)
}
}
