% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSMOTE.R
\name{newSample}
\alias{newSample}
\title{Auxiliary function used by MLSMOTE. Creates a synthetic sample based on values of attributes and labels of its neighbors}
\usage{
newSample(sample, refNeigh, neighbors, D)
}
\arguments{
\item{sample}{Sample we are using as "template"}

\item{refNeigh}{Reference neighbor}

\item{neighbors}{Neighbors to take into account}

\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}
}
\value{
A synthetic sample derived from the one passed as a parameter and its neighbors
}
\description{
Auxiliary function used by MLSMOTE. Creates a synthetic sample based on values of attributes and labels of its neighbors
}
\examples{
newSample(25,28,c(15,28,62),bibtex)
}
