% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setParallel}
\alias{setParallel}
\title{Enable/Disable parallel computing}
\usage{
setParallel(beParallel)
}
\arguments{
\item{beParallel}{A boolean indicating if parallel computing is to be enabled (TRUE) or disabled (FALSE)}
}
\value{
No return value, called in order to enable parallel computing
}
\description{
Enable/Disable parallel computing
}
\examples{
setParallel(TRUE)

}
