% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vdm}
\alias{vdm}
\title{Auxiliary function used to calculate the Value Difference Metric (VDM) between two instances considering their non numeric attributes}
\usage{
vdm(D, sample, y, label, tableVDM = NULL)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{sample}{Index of the first sample}

\item{y}{Index of the second sample}

\item{label}{Label that will be considered in calculations}

\item{tableVDM}{Dataframe object containing previous calculations for faster processing. If it is empty, the algorithm will be slower}
}
\value{
A value for the distance
}
\description{
Auxiliary function used to calculate the Value Difference Metric (VDM) between two instances considering their non numeric attributes
}
