% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistances.R
\name{calculateDistances}
\alias{calculateDistances}
\title{Auxiliary function used by to calculate the distances between an instance and the ones with a specific active label}
\usage{
calculateDistances(sample, rest, label, D)
}
\arguments{
\item{sample}{Index of the sample whose distances to other samples we want to know}

\item{rest}{Indexes of the samples to which we will calculate the distance}

\item{label}{Label that must be active}

\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}
}
\value{
A list with the distance to the rest of samples
}
\description{
Auxiliary function used by to calculate the distances between an instance and the ones with a specific active label
}
\examples{
\dontrun{
library(mldr)
calculateDistances(25,c(75,65,89,23),300,bibtex)
}
}
