% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllNeighbors}
\alias{getAllNeighbors}
\title{Auxiliary function used by MLSOL and MLUL. Computes the kNN of every instance in a dataset}
\usage{
getAllNeighbors(D, d, tableVDM = NULL)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}

\item{tableVDM}{Dataframe object containing previous calculations for faster processing. If it is empty, the algorithm will be slower}
}
\value{
A list of vectors with the indexes of the neighbors for each instance
}
\description{
Auxiliary function used by MLSOL and MLUL. Computes the kNN of every instance in a dataset
}
