% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getV}
\alias{getV}
\title{Auxiliary function used by MLUL. It calculates, for each instance, how important it is in the dataset}
\usage{
getV(w, u)
}
\arguments{
\item{w}{List of weights for each instance}

\item{u}{List of influences in reverse neighbors for each instance}
}
\value{
A list with the values of importance of each instance in the dataset
}
\description{
Auxiliary function used by MLUL. It calculates, for each instance, how important it is in the dataset
}
