% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initTypes}
\alias{initTypes}
\title{Auxiliary function used by MLSOL. Categorizes each pair instance-label of the dataset with a type}
\usage{
initTypes(C, neighbors, k, minoritary, D, d)
}
\arguments{
\item{C}{List of vectors with one value for each pair instance-label}

\item{neighbors}{Structure with the k nearest neighbors of each instance of the dataset}

\item{k}{Number of neighbors to be considered for each instance}

\item{minoritary}{Vector with the minoritary value of each label (normally, 1)}

\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{d}{Vector with the instances of the dataset which have one or more label active (ideally, all of them)}
}
\value{
A synthetic sample derived from the one passed as a parameter and its neighbors
}
\description{
Auxiliary function used by MLSOL. Categorizes each pair instance-label of the dataset with a type
}
