% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLSMOTE.R
\name{MLSMOTE}
\alias{MLSMOTE}
\title{Synthetic oversampling of multilabel instances (MLSMOTE)}
\source{
Charte, F., Rivera, A. J., del Jesus, M. J., & Herrera, F. (2015). MLSMOTE: Approaching imbalanced multilabel learning through synthetic instance generation. Knowledge-Based Systems, 89, 385-397.
}
\usage{
MLSMOTE(D, k, strategy = "ranking", tableVDM = NULL)
}
\arguments{
\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}

\item{k}{Number of neighbors to be considered when creating a synthetic instance}

\item{strategy}{Strategy for choosing the synthetic labels. Possible values: "union", "intersection" and "ranking" (default)}

\item{tableVDM}{Dataframe object containing previous calculations for faster processing. If it is empty, the algorithm will be slower}
}
\value{
A mld object containing the preprocessed multilabel dataset
}
\description{
This function implements the MLSMOTE algorithm. It is a preprocessing algorithm for imbalanced multilabel datasets,
whose aim is to identify instances with minoritary labels, and generate synthetic instances based on their neighbor instances.
}
