% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newSample}
\alias{newSample}
\title{Auxiliary function used by MLSMOTE. Creates a synthetic sample based on values of attributes and labels of its neighbors}
\usage{
newSample(seedInstance, refNeigh, neighbors, strategy, D)
}
\arguments{
\item{seedInstance}{Sample we are using as "template"}

\item{refNeigh}{Reference neighbor}

\item{neighbors}{Neighbors to take into account}

\item{strategy}{Strategy for choosing the synthetic labels: union, intersection or ranking}

\item{D}{mld \code{mldr} object with the multilabel dataset to preprocess}
}
\value{
A synthetic sample derived from the one passed as a parameter and its neighbors
}
\description{
Auxiliary function used by MLSMOTE. Creates a synthetic sample based on values of attributes and labels of its neighbors
}
