% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setNumCores}
\alias{setNumCores}
\title{Set the number of cores available for parallel computing}
\usage{
setNumCores(n)
}
\arguments{
\item{n}{The new value for the number of cores}
}
\value{
No return value, called in order to change the number of cores
}
\description{
Set the number of cores available for parallel computing
}
\examples{
\donttest{
setNumCores(8)
}
}
